Move ochre dynamic control to a new script
